precompile copycat_engine.css when deploying to heroku
The problem before was that "initialize_on_precompile" was set to false. That variable causes only part of the rails env to be run -- specifically, only the initializers under the group 'assets.' Even though the initializer was getting created, it wasn't getting run because it's group was set to some default value. Adding it to the :all group causes it to always be run, and copycat_engine.css is always precompiled.